pyproject.toml
[project]
name = "celery-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "celery[redis]>=5.6.2",
    "flask>=3.1.3",
    "psycopg[binary,pool]>=3.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
